Move type definitions into _internal
Also rename from `_types` (generic, unclear) to `type_definitions`
(still generic, hopefully less unclear).
