+++ src/globus_sdk/globus_app/protocols.py
@@ -4,7 +4,7 @@ import typing as t
 
 if t.TYPE_CHECKING:
     from globus_sdk import AuthLoginClient, IDTokenDecoder
-    from globus_sdk._types import UUIDLike
+    from globus_sdk._internal.type_definitions import UUIDLike
     from globus_sdk.login_flows import LoginFlowManager
     from globus_sdk.token_storage import TokenStorage, TokenValidationError
