+++ src/globus_sdk/scopes/data/flows.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import typing as t
 
-from globus_sdk._types import UUIDLike
+from globus_sdk._internal.type_definitions import UUIDLike
 
 from ..collection import (
     DynamicScopeCollection,
