+++ src/globus_sdk/services/auth/client/confidential_client.py
@@ -5,8 +5,8 @@ import typing as t
 
 from globus_sdk import exc
 from globus_sdk._internal.remarshal import commajoin, strseq_iter, strseq_listify
+from globus_sdk._internal.type_definitions import UUIDLike
 from globus_sdk._missing import MISSING, MissingType
-from globus_sdk._types import UUIDLike
 from globus_sdk.authorizers import BasicAuthorizer
 from globus_sdk.response import GlobusHTTPResponse
 from globus_sdk.scopes import Scope
