+++ src/globus_sdk/services/compute/data.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
 
+from globus_sdk._internal.type_definitions import UUIDLike
 from globus_sdk._missing import MISSING, MissingType
 from globus_sdk._payload import GlobusPayload
-from globus_sdk._types import UUIDLike
 from globus_sdk.exc import warn_deprecated
