+++ src/globus_sdk/services/gcs/client.py
@@ -6,9 +6,9 @@ import uuid
 from globus_sdk import client, exc, paging, response
 from globus_sdk._internal.classprop import classproperty
 from globus_sdk._internal.remarshal import commajoin
+from globus_sdk._internal.type_definitions import UUIDLike
 from globus_sdk._internal.utils import slash_join
 from globus_sdk._missing import MISSING, MissingType
-from globus_sdk._types import UUIDLike
 from globus_sdk.authorizers import GlobusAuthorizer
 from globus_sdk.globus_app import GlobusApp
 from globus_sdk.scopes import GCSCollectionScopes, GCSEndpointScopes, Scope
