+++ src/globus_sdk/services/gcs/data/role.py
@@ -2,9 +2,9 @@ from __future__ import annotations
 
 import typing as t
 
+from globus_sdk._internal.type_definitions import UUIDLike
 from globus_sdk._missing import MISSING, MissingType
 from globus_sdk._payload import GlobusPayload
-from globus_sdk._types import UUIDLike
 
 
 class GCSRoleDocument(GlobusPayload):
