+++ src/globus_sdk/services/groups/data.py
@@ -4,9 +4,9 @@ import enum
 import typing as t
 
 from globus_sdk._internal.remarshal import strseq_iter
+from globus_sdk._internal.type_definitions import UUIDLike
 from globus_sdk._missing import MISSING, MissingType
 from globus_sdk._payload import GlobusPayload
-from globus_sdk._types import UUIDLike
 
 T = t.TypeVar("T")
