+++ src/globus_sdk/services/transfer/client.py
@@ -8,8 +8,8 @@ import uuid
 from globus_sdk import client, exc, paging, response
 from globus_sdk._internal import guards
 from globus_sdk._internal.remarshal import commajoin
+from globus_sdk._internal.type_definitions import DateLike, IntLike, UUIDLike
 from globus_sdk._missing import MISSING, MissingType
-from globus_sdk._types import DateLike, IntLike, UUIDLike
 from globus_sdk.scopes import GCSCollectionScopes, Scope, TransferScopes
 
 from .data import DeleteData, TransferData
