+++ src/globus_sdk/services/transfer/data/delete_data.py
@@ -5,9 +5,9 @@ import logging
 import typing as t
 
 from globus_sdk._internal.remarshal import stringify
+from globus_sdk._internal.type_definitions import UUIDLike
 from globus_sdk._missing import MISSING, MissingType
 from globus_sdk._payload import GlobusPayload
-from globus_sdk._types import UUIDLike
 
 log = logging.getLogger(__name__)
